+++ apps/customize_menu.sh
@@ -11,6 +11,7 @@ customize_menu() {
             "Shell" \
             "FastFetch" \
             "Code Editor" \
+            "Dock Layout" \
             "Spicetify" \
             "---" \
             "Wallpaper" \
@@ -21,8 +22,9 @@ customize_menu() {
             2) source "$MACRIFT_DIR/apps/terminal.sh" && shell_menu ;;
             3) source "$MACRIFT_DIR/apps/terminal.sh" && fastfetch_menu ;;
             4) source "$MACRIFT_DIR/apps/editor.sh" && editor_menu ;;
-            5) source "$MACRIFT_DIR/apps/spicetify.sh" && restore_marketplace ;;
-            6) source "$MACRIFT_DIR/apps/wallpaper.sh" && wallpaper_menu ;;
+            5) source "$MACRIFT_DIR/apps/dock_layout.sh" && dock_layout_menu ;;
+            6) source "$MACRIFT_DIR/apps/spicetify.sh" && restore_marketplace ;;
+            7) source "$MACRIFT_DIR/apps/wallpaper.sh" && wallpaper_menu ;;
             0) return ;;
             *) ;;
         esac